♻️ [Refactor]: chat folder template architecture
- chat에서만 사용되는 컴포넌트가 있어 templates/chats에서 관리
- chat page에서 필요한 component는 templates/chat/index.tsx에서 export
