+++ src/components/templates/chats/index.tsx
@@ -1,0 +1,4 @@
+import ChatList from '@/components/templates/chats/ChatList';
+import ChatRoom from '@/components/templates/chats/ChatRoom';
+
+export { ChatList, ChatRoom };
